\docType{data}
\name{bison}
\alias{bison}
\title{Bison GPS coordinates}
\format{Data frame}
\description{
  GPS location information for one American bison
  (\emph{Bison bison}) at The Nature Conservancy Tallgrass
  Prairie Preserve, Jan 31 - Feb 5, 2011. Information
  includes timestamp, latitude, and longitude.
}
\keyword{datasets}

