\docType{package}
\name{okmesonet-package}
\alias{okmesonet}
\alias{okmesonet-package}
\title{Oklahoma Mesonet climate data}
\description{
  The \href{http://www.mesonet.org/}{Oklahoma Mesonet} is a
  network of automated climate monitoring stations
  throughout Oklahoma, USA. Data collection began January
  01, 1994; as of February 2012, there are 120 active
  stations. Measurements are recorded every five minutes
  and sent to a central facility for verification and
  quality control by the Oklahoma Climatological Survey.
}

