\docType{data}
\name{bison}
\alias{bison}
\title{Bison locations}
\format{Data frame}
\description{
  Location information for one American bison (\emph{Bison
  bison}) at The Nature Conservancy Tallgrass Prairie
  Preserve, Jan 31 - Feb 5, 2011. Information includes
  time, latitude, and longitude.
}
\keyword{datasets}

