\docType{package}
\name{okmesonet-package}
\alias{okmesonet}
\alias{okmesonet-package}
\title{Oklahoma Mesonet climatological data}
\description{
  The \href{http://www.mesonet.org/}{Oklahoma Mesonet} is a
  network of automated climate monitoring stations
  throughout Oklahoma, USA. Data collection began January
  01, 1994; as of February 2012, there are 120 active
  stations. Measurements are recorded every five minutes
  and sent to a central facility for verification and
  quality control by the Oklahoma Climatological Survey.
}
\details{
  Data access may be restricted by organization and/or
  location. Please refer to and follow policies found
  within the
  \href{http://www.mesonet.org/index.php/site/about/data_access_and_pricing}{Oklahoma
  Mesonet Data Access Policy}. The authors and maintainers
  of \code{okmesonet} assume no responsibility for the use
  or misuse of \code{okmesonet}.
}

