\docType{package}
\name{okmesonet-package}
\alias{okmesonet}
\alias{okmesonet-package}
\title{Access available Oklahoma Mesonet climate data}
\description{
  The \href{http://www.mesonet.org/}{Oklahoma Mesonet} is a
  network of automated climate monitoring stations
  throughout the state of Oklahoma, USA; data collection
  began on Jan 01, 1994. As of February 2012, there are 120
  active stations, with an additional 15 stations that have
  been decommissioned (but with available data).
  Measurements are recorded every five minutes and sent to
  a central facility for verification and quality control
  by the Oklahoma Climatological Survey.
}

