\name{updatestn}
\alias{updatestn}
\title{Update list of Oklahoma Mesonet stations}
\usage{
  updatestn()
}
\description{
  The list of Oklahoma Mesonet stations was updated November 28, 2014. 
  Use \code{updatestn} to manually update the list.
}
\examples{
\dontrun{
## Update Oklahoma Mesonet station list
updatestn()
}
}
\seealso{
  \code{\link{okstations}}
}

